# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minut Booking API",
    "version": "1.0.0",
    "description": "REST API for managing rental units and guest reservations. All endpoints are versioned under /v1. Reservations use half-open date ranges [checkIn, checkOut) and overlapping reservations for the same unit are rejected with 409."
  },
  "servers": [
    {
      "url": "http://localhost:5006"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "RentalUnitList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalUnit"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "RentalUnit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "addressLine1": {
            "type": [
              "string",
              "null"
            ]
          },
          "addressLine2": {
            "type": [
              "string",
              "null"
            ]
          },
          "city": {
            "type": [
              "string",
              "null"
            ]
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer",
            "description": "Send back unchanged when updating"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "addressLine1",
          "addressLine2",
          "city",
          "postalCode",
          "country",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "RentalUnitEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RentalUnit"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateRentalUnit": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Sunny loft in Södermalm"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "addressLine1": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "example": "Götgatan 12"
          },
          "addressLine2": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "city": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100,
            "example": "Stockholm"
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20,
            "example": "116 46"
          },
          "country": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100,
            "example": "SE"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRentalUnit": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Sunny loft in Södermalm"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "addressLine1": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "example": "Götgatan 12"
          },
          "addressLine2": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "city": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100,
            "example": "Stockholm"
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20,
            "example": "116 46"
          },
          "country": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100,
            "example": "SE"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "description": "The version you last read. Mismatch → 409 VERSION_CONFLICT"
          }
        },
        "required": [
          "version"
        ]
      },
      "ReservationList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rentalUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "rentalUnit": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "description": "Denormalised for list views"
          },
          "guestName": {
            "type": "string"
          },
          "guestEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "guestCount": {
            "type": "integer"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01",
            "description": "Exclusive — the guest leaves this day"
          },
          "nights": {
            "type": "integer"
          },
          "version": {
            "type": "integer",
            "description": "Send back unchanged when updating"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "rentalUnitId",
          "rentalUnit",
          "guestName",
          "guestEmail",
          "guestCount",
          "notes",
          "checkIn",
          "checkOut",
          "nights",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "RevisionList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Revision"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "Revision": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer"
          },
          "changeType": {
            "type": "string",
            "enum": [
              "update",
              "delete"
            ]
          },
          "snapshot": {
            "type": "object",
            "additionalProperties": {},
            "description": "The row as it was before the change"
          },
          "changedBy": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "version",
          "changeType",
          "snapshot",
          "changedBy",
          "changedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/rental-units": {
      "get": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List rental units",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "description": "1-based page number"
            },
            "required": false,
            "description": "1-based page number",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Items per page (max 100)"
            },
            "required": false,
            "description": "Items per page (max 100)",
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 200,
              "description": "Case-insensitive name/city search"
            },
            "required": false,
            "description": "Case-insensitive name/city search",
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalUnitList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a rental unit",
        "description": "Requires the admin role.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentalUnit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalUnitEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rental-units/{id}": {
      "get": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a rental unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalUnitEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a rental unit",
        "description": "Partial update. `version` must match the current version (optimistic locking).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalUnit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalUnitEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (overlap, stale version, or dependent records)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a rental unit",
        "description": "Soft delete. Fails with 409 while the unit still has reservations.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (overlap, stale version, or dependent records)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rental-units/{id}/reservations": {
      "get": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List reservations of a rental unit",
        "description": "Optionally limited to those intersecting the half-open period [from, to).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "description": "1-based page number"
            },
            "required": false,
            "description": "1-based page number",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Items per page (max 100)"
            },
            "required": false,
            "description": "Items per page (max 100)",
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays ending after this date"
            },
            "required": false,
            "description": "Include stays ending after this date",
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays starting before this date"
            },
            "required": false,
            "description": "Include stays starting before this date",
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "checkIn",
                "-checkIn",
                "createdAt",
                "-createdAt"
              ],
              "default": "checkIn"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rental-units/{id}/history": {
      "get": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change history of a rental unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}